feat!: move config commands to top level
