+++ cmd/soroban-cli/src/commands/contract/invoke.rs
@@ -185,7 +185,7 @@ impl Cmd {
                 if let Some(mut val) = matches_.get_raw(&name) {
                     let mut s = val.next().unwrap().to_string_lossy().to_string();
                     if matches!(i.type_, ScSpecTypeDef::Address) {
-                        let cmd = crate::commands::config::identity::address::Cmd {
+                        let cmd = crate::commands::identity::address::Cmd {
                             name: Some(s.clone()),
                             hd_path: Some(0),
                             locator: self.config.locator.clone(),
