+++ cmd/soroban-cli/src/commands/identity/add.rs
@@ -1,8 +1,12 @@
-use super::super::locator;
 use clap::command;
+
+use super::super::config::{locator, secret};
 
 #[derive(thiserror::Error, Debug)]
 pub enum Error {
+    #[error(transparent)]
+    Secret(#[from] secret::Error),
+
     #[error(transparent)]
     Config(#[from] locator::Error),
 }
@@ -10,13 +14,20 @@ pub enum Error {
 #[derive(Debug, clap::Parser, Clone)]
 #[group(skip)]
 pub struct Cmd {
+    /// Name of identity
+    pub name: String,
+
+    #[command(flatten)]
+    pub secrets: secret::Args,
+
     #[command(flatten)]
     pub config_locator: locator::Args,
 }
 
 impl Cmd {
     pub fn run(&self) -> Result<(), Error> {
-        println!("{}", self.config_locator.list_identities()?.join("\n"));
-        Ok(())
+        Ok(self
+            .config_locator
+            .write_identity(&self.name, &self.secrets.read_secret()?)?)
     }
 }
