+++ cmd/soroban-cli/src/commands/identity/ls.rs
@@ -1,11 +1,9 @@
-use super::{super::secret, locator};
 use clap::command;
+
+use super::super::config::locator;
 
 #[derive(thiserror::Error, Debug)]
 pub enum Error {
-    #[error(transparent)]
-    Secret(#[from] secret::Error),
-
     #[error(transparent)]
     Config(#[from] locator::Error),
 }
@@ -13,20 +11,13 @@ pub enum Error {
 #[derive(Debug, clap::Parser, Clone)]
 #[group(skip)]
 pub struct Cmd {
-    /// Name of network
-    pub name: String,
-
-    #[command(flatten)]
-    pub network: super::Network,
-
     #[command(flatten)]
     pub config_locator: locator::Args,
 }
 
 impl Cmd {
     pub fn run(&self) -> Result<(), Error> {
-        Ok(self
-            .config_locator
-            .write_network(&self.name, &self.network)?)
+        println!("{}", self.config_locator.list_identities()?.join("\n"));
+        Ok(())
     }
 }
