+++ cmd/soroban-cli/src/commands/identity/show.rs
@@ -1,5 +1,6 @@
-use super::super::{locator, secret};
 use clap::arg;
+
+use super::super::config::{locator, secret};
 
 #[derive(thiserror::Error, Debug)]
 pub enum Error {
