+++ cmd/soroban-cli/src/commands/network/ls.rs
