+++ cmd/soroban-cli/src/commands/network/rm.rs
